requirements.txt
Flask
pypdf==3.17.4
python-docx
pdf2docx
docx2pdf
gunicorn 